Updated changes of march

+++ quotation/views.py
@@ -51,8 +51,8 @@ def quotation_generate(request):
                                                        part_quantity=part.get(
                                                            'part_quantity'),
                                                        created_by=request.user)
-                    part_total_cost += (int(obj.price)
-                                        * int(obj.part_quantity))
+                    part_total_cost += (float(obj.price)
+                                        * float(obj.part_quantity))
                     part_obj.append(obj)
 
         quotation_obj.parts.add(*part_obj)
